fixing #20 from int to long and updating nuget packages with docs

+++ src/Bet.Extensions.Http.MessageHandlers/Authorize/AuthorizeTokenResponse.cs
@@ -28,7 +28,7 @@ namespace Bet.Extensions.Http.MessageHandlers.Authorize
         public string TokenType { get; set; } = string.Empty;
 
         [JsonProperty("expires_in")]
-        public int ExpiresInSeconds { get; set; }
+        public long ExpiresInSeconds { get; set; }
 
         public TimeSpan ExpiresIn => TimeSpan.FromSeconds(ExpiresInSeconds);
     }
